% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{oq_pal_cat}
\alias{oq_pal_cat}
\title{The categorical color palette}
\usage{
oq_pal_cat(reverse)
}
\description{
The categorical color palette
}
\examples{
library(scales)
scales::show_col(oq_pal_cat()(9))
}
